--- a/go.mod
+++ b/go.mod
@@ -4,5 +4,6 @@ go 1.12
 
 require (
 	github.com/davecgh/go-spew v1.1.1
+	github.com/json-iterator/go v1.1.7
 	gopkg.in/yaml.v2 v2.2.2
 )