add a couple of json interator configurations:

1) jsonIterator should be compatible with the old encoding/json
there is a breaking change here because the "JSONOpt" is now
a no-op, although the options in encoding/json are not very useful.

2) caseSensitiveStrictJSONIterator is compatible with
econding/json WRT to DisallowUnknownFields, but it also has
CaseSensitive. This is a breaking change for strict unmarshal users
that previously allowed case-insensitive fields.

Something else to note is that json-iter does not seem to tolerate
field keys called "true", which the one from stdlib does.
Unit tests had to be adapted because of that.

Other changes:
- add a new public method UnmarshalWithConfig that allows passing
a json-iter configuration
- update/add unit tests
- update go.mod/sum
- remove yaml_go110*.go
These files had the purpose to handle DisallowUnknownFields for
json.Decoder which is no longer needed.

/kind feature
/priority important-longterm

fixes https://github.com/kubernetes-sigs/yaml/issues/17
fixes https://github.com/kubernetes-sigs/yaml/issues/15

/assign @dims 
/cc @sttts @errordeveloper @stealthybox
/hold