@nikhita 2019-06-03
> @smarterclayton, didn't you see performance regressions in json-iter marshaling?

Ref: https://github.com/kubernetes/kubernetes/pull/70574#issuecomment-476370725, https://github.com/kubernetes/kubernetes/pull/63242

The second PR referenced above did have high allocations:

Running `go test ./vendor/k8s.io/apiserver/pkg/endpoints/ -test.bench='Benchmark(Get|WatchHTTP)' -test.run=XXX -benchmem -test.memprofile=/tmp/mem2 -test.benchtime 5s`

On k/k master right now:

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/endpoints
BenchmarkGet-4              	   30000	    256834 ns/op	   15794 B/op	     127 allocs/op
BenchmarkWatchHTTP-4        	  100000	     92040 ns/op	    1523 B/op	      24 allocs/op
BenchmarkWatchHTTP_UTF8-4   	  100000	     96523 ns/op	    1829 B/op	      26 allocs/op
PASS
ok  	k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/endpoints	31.248s
```

k/k right now with changes in the second PR (https://github.com/kubernetes/kubernetes/pull/63242):
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/endpoints
BenchmarkGet-4                     10000	    691435 ns/op	   19122 B/op	     163 allocs/op
BenchmarkWatchHTTP-4        	  200000	     48193 ns/op	    5495 B/op	      64 allocs/op
BenchmarkWatchHTTP_UTF8-4   	  200000	     53738 ns/op	    5909 B/op	      69 allocs/op
PASS
ok  	k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/endpoints	28.604s
```

k/k with changes in this PR:

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/endpoints
BenchmarkGet-4              	   50000	    139204 ns/op	   15793 B/op	     127 allocs/op
BenchmarkWatchHTTP-4        	  200000	     56627 ns/op	    1522 B/op	      24 allocs/op
BenchmarkWatchHTTP_UTF8-4   	  100000	     62922 ns/op	    1829 B/op	      26 allocs/op
PASS
ok  	k8s.io/kubernetes/vendor/k8s.io/apiserver/pkg/endpoints	27.378s
```

So it should probably be fine...