@nikhita 2019-06-04
> You should also run the serialization benchmarks in pkg/api/testing/serialization_test.go to get a before and after round trip time (that tests several variants of objects, whereas BenchmarkGet tests a single known object).

Running existing benchmarks (before and after) yielded similar results...but that isn't surprising because none of the benchmarks actually look for yaml marshaling and unmarshaling.

Created https://github.com/kubernetes/kubernetes/pull/78688 for adding benchmarks for yaml marshaling and unmarshaling as well. Using these benchmarks, incorporating changes in this PR into k/k yield higher allocations... :(

Running `go test k8s.io/kubernetes/pkg/api/testing/ -test.bench='BenchmarkEncodeYAMLMarshal|BenchmarkDecodeIntoYAML' -test.run=XXX -benchmem -test.memprofile=/tmp/mem2 -test.benchtime 5s`

**Before**:

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/api/testing
BenchmarkEncodeYAMLMarshal-4   	   10000	    781873 ns/op	  256102 B/op	    1438 allocs/op
BenchmarkDecodeIntoYAML-4      	   10000	    627891 ns/op	  167788 B/op	    1485 allocs/op
PASS
ok  	k8s.io/kubernetes/pkg/api/testing	14.354s
```

**After**:

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/api/testing
BenchmarkEncodeYAMLMarshal-4   	   10000	    867119 ns/op	  260776 B/op	    1443 allocs/op
BenchmarkDecodeIntoYAML-4      	   10000	    692040 ns/op	  181813 B/op	    1609 allocs/op
PASS
ok  	k8s.io/kubernetes/pkg/api/testing	15.842s
```