@smarterclayton 2019-06-04
What is the underlying goal of this PR?  Use json-iterator because <xxx>?  Consistency with Kube is important, but it's not all.  Was there a discussion about the why?  I don't see many comments in #17