@neolit123 2019-06-04
> What is the underlying goal of this PR? Use json-iterator because ? Consistency with Kube is important, but it's not all. Was there a discussion about the why? I don't see many comments in #17

main driver was discussed in https://github.com/kubernetes-sigs/yaml/issues/15
the current library's strict unmarshal is not case-sensitive.