@neolit123 2019-07-26
here is an example of this problem in the wild:
https://github.com/kubernetes/kubernetes/issues/80582

kubeadm has strict unmarshaling to show warnings and then perform regular unmarshall (two pass).
however the upper case `ClusterConfiguration` field was not caught; should have been `clusterConfiguration`.