@neolit123 2019-08-04
so going back to the OP and the discussion there are at lease 3 concerns here:

- benchmark results
the overhead after the change is not high and fixing the bug for StrictUnmarshal users can justify the small decrease in performance. on the other hand there aren't that many StrictUnmarshal users out there, so the decrease in performance is probably making a more apparent tilt of the scale.

- breaking change 1:
> This is a breaking change for strict unmarshal users
that previously allowed case-insensitive fields.

we enabled strict unmarshal in this library not so long ago, and given strict unmarshal is optional it seems that this might not affect that many users. but this is still probably going to break someone, somewhere once they move to the latest SHA of this repository.

- breaking change 2:
> there is a breaking change here because the "JSONOpt" is now
a no-op, although the options in encoding/json are not very useful.

not sure if this will break anyone, probably not.

---
i will update this PR and leave it for a couple of more weeks.