@neolit123 2019-08-15
clearing some pending PRs from my backlog.
if someone feels like they want to take this over i will leave the branch.
thanks!